+++ core/lib/circuit/src/witness/forced_exit.rs
@@ -75,7 +75,7 @@ impl Witness for ForcedExitWitness<Bn256> {
                 .to_u128()
                 .unwrap(),
             fee: forced_exit.tx.fee.to_u128().unwrap(),
-            token: *forced_exit.tx.token as u32,
+            token: *forced_exit.tx.token,
             initiator_account_address: *forced_exit.tx.initiator_account_id,
             target_account_address: *forced_exit.target_account_id,
             target_account_eth_address: eth_address_to_fr(&forced_exit.tx.target),
